bot/requirements.txt

python-telegram-bot
requests

